refactor: rename balances state files to plural to be named as the exported var
