+++ packages/shell/src/State/ShellContext.ts
@@ -1,6 +1,6 @@
 import { Dispatch, createContext } from "react";
 import { AdapterAction } from "../Adapter/State/AdapterActions";
-import { BalanceAction } from "../Balances/State/BalanceActions";
+import { BalanceAction } from "../Balances/State/BalancesActions";
 import { getShellInitialState, ShellState } from "./ShellState";
 
 export type ShellAction = AdapterAction | BalanceAction;
